+++ test/Spec.hs
@@ -1,13 +1,15 @@
 {-# LANGUAGE ImportQualifiedPost #-}
-import Test.Tasty ( TestTree, defaultMain, testGroup )
-import Test.Tasty.HUnit ( testCase, (@?=) )
+{-# OPTIONS_GHC -Wno-incomplete-patterns #-}
+
+import Lib2
+import Lib3
+import Test.Tasty (TestTree, defaultMain, testGroup)
+import Test.Tasty.HUnit (testCase, (@?=))
 import Test.Tasty.QuickCheck as QC
 
-import Data.List
-import Data.Ord
 
-import Lib1 qualified
-import Lib2 qualified
+--import Debug.Trace (trace)
+--trace ("Parsed name: " ++ show name ++ ", Remaining: " ++ show rest1) $
 
 main :: IO ()
 main = defaultMain tests
@@ -16,18 +18,23 @@ tests :: TestTree
 tests = testGroup "Tests" [unitTests, propertyTests]
 
 unitTests :: TestTree
-unitTests = testGroup "Lib1 tests"
-  [ testCase "List of completions is not empty" $
-      null Lib1.completions @?= False,
-    testCase "Parsing case 1 - give a better name" $
-      Lib2.parseQuery "" @?= (Left "Some error message"),
-    testCase "Parsing case 2 - give a better name" $
-      Lib2.parseQuery "o" @?= (Left "Some error message")
-  ]
+unitTests =
+  testGroup
+    "Lib2 tests"
+    [ testCase "Parsing case 1 - give a better name" $
+        Lib2.parseQuery "" @?= Left "Command cannot be empty",
+      testCase "Parsing case 2 - give a better name" $
+        Lib2.parseQuery "o" @?= Left "Invalid command"
+    ]
+
 
 propertyTests :: TestTree
-propertyTests = testGroup "some meaningful name"
-  [
-    QC.testProperty "sort == sort . reverse" $
-      \list -> sort (list :: [Int]) == sort (reverse list)
-  ]
+propertyTests =
+  testGroup
+    "Property tests"
+    [ QC.testProperty "parseQuery . renderQuery == Right query" $
+        \query ->
+          --trace ("query: " ++ show query) $
+          --trace ("Lib2.parseQuery (Lib3.renderQuery query): " ++ show (Lib2.parseQuery (Lib3.renderQuery query))) $
+          Lib2.parseQuery (Lib3.renderQuery query) == Right query
+    ]